+++ Voltyks.Core/DTOs/Store/Admin/AdminReservationDto.cs
@@ -19,8 +19,8 @@ namespace Voltyks.Core.DTOs.Store.Admin
 
         // Order Details
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal TotalPrice { get; set; }
+        public int UnitPrice { get; set; }
+        public int TotalPrice { get; set; }
         public string Currency { get; set; } = "EGP";
         public string Status { get; set; } = string.Empty;
